fix: login redirect looping
the portal cookie set in this program does not assign a domain for it,
so it will be assigned to the root domain by default.

by this, it has a higher priority than the cookie assigned to the
subdomain. when the time in front of the portal cookie expires, the
portal will try to renew the time by setting a new cookie with a new
time for you.

due tothe cause explained earlier, the portal will set a new cookie
with the domain 'portal.ncu.edu.tw'. next time when the program is
redirected to the login page, it reads the cookie with the root
domain which the time not updated yet.

then the page will try to update it again, causing an infinite looping.

This commit adds a domain assignment to the portal cookie to fix the
problem.