feat: discord notification
add a feature to send the task status message to a Discord channel.

the notifier can add the other notification channel, just implement the
channel class and include a notify function.

then, add the channel class object to the notifier, the notify function
will be called when there has a message needs to be sent.